Pass container envars to eyeos-auth so dovecot auth works
Dovecot executes a script to authenticate users/password. This script
uses our eyeos-auth, that gets the public key (to validate signatures)
from the environment. But dovecot nullifies the environment before
executing the script, so eyeos-auth cannot retrieve the public key from
the environment. So by saving the environment to a file at container
startup and loading it in the checkpassword script (the one that dovecot
executes without environment) we can set the environment accordingly for
eyeos-auth to work.